requirements.txt
pandas>=2.0
numpy>=1.24
pyarrow>=14.0
pyyaml>=6.0
tqdm>=4.65
loguru>=0.7
requests>=2.31
pydantic>=2.5
readability-lxml>=0.8.1
lxml>=5.1.0
beautifulsoup4>=4.12

# Data providers
tushare>=1.2
akshare>=1.12
curl_cffi>=0.6
jsonschema>=4.20
duckdb>=0.10
fastapi>=0.110
uvicorn>=0.24
httpx>=0.26
pytest>=7.0
tzdata>=2023.3
